+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
-import './widgets/user_transaction.dart';
+import './widgets/transactionList.dart';
+import './widgets/new_transaction.dart';
+import 'models/Transaction.dart';
 
 void main() => runApp(MyApp());
 
@@ -17,15 +19,43 @@ class MyApp extends StatelessWidget {
   }
 }
 
-class MyHomePage extends StatelessWidget {
-  // final List<Transaction> transactions = [
-  // ];
-
-  // String titleImput;
+class MyHomePage extends StatefulWidget {
+  // String titleInput;
   // String amountInput;
+  @override
+  _MyHomePageState createState() => _MyHomePageState();
+}
 
-  final titleController = TextEditingController();
-  final amountController = TextEditingController();
+class _MyHomePageState extends State<MyHomePage> {
+
+  final List<Transaction> _userTransaction = [
+    Transaction(id: 't1', tile: 'Nike', amount: 69.99, date: DateTime.now()),
+  ];
+
+  void _addNewTransaction(String title, double amount) {
+    
+    final newTx = Transaction(
+        tile: title,
+        amount: amount,
+        date: DateTime.now(),
+        id: DateTime.now().toString());
+       
+        
+        setState(() {
+          _userTransaction.add(newTx);
+        });
+  }
+
+  void _startAddNewTransaction(BuildContext ctx){
+   showModalBottomSheet(context: ctx, builder: (_) {
+
+       return GestureDetector(
+         onTap: (){},
+         child: NewTransaction(addtransation: _addNewTransaction) ,
+         behavior: HitTestBehavior.opaque ,
+       );
+   });
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -33,29 +63,33 @@ class MyHomePage extends StatelessWidget {
       appBar: AppBar(
         title: Text('Flutter Demo '),
         actions: <Widget>[
-          IconButton(icon: Icon(Icons.add), onPressed: () {} ,)
+          IconButton(
+            icon: Icon(Icons.add),
+            onPressed: () => _startAddNewTransaction(context) ,
+          )
         ],
       ),
-      body:  SingleChildScrollView(
-              child: Column(
-              //mainAxisAlignment: MainAxisAlignment.start,
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: <Widget>[
-                Container(
-                  width: double.infinity,
-                  child: Card(
-                    color: Colors.blue,
-                    child: Text('CHART'),
-                    elevation: 5,
-                  ),
+      body: SingleChildScrollView(
+        child: Column(
+            //mainAxisAlignment: MainAxisAlignment.start,
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            children: <Widget>[
+              Container(
+                width: double.infinity,
+                child: Card(
+                  color: Colors.blue,
+                  child: Text('CHART'),
+                  elevation: 5,
                 ),
-                UserTransaction()
-              ]),
+              ),
+              TransactionList(transactions: _userTransaction)
+            ]),
       ),
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat ,
-      floatingActionButton: FloatingActionButton(child: Icon(Icons.add) , onPressed: (){
-         
-      } ,),
+      floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
+      floatingActionButton: FloatingActionButton(
+        child: Icon(Icons.add),
+        onPressed: () => _startAddNewTransaction(context) ,
+      ),
     );
   }
 }
